requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
pypdfium2>=0.23.0
Pillow>=10.0.0
openpyxl>=3.1.0
requests>=2.31.0
pydantic>=2.0.0
httpx>=0.27.0
openai>=1.0.0
